Rename UserResponse as UserRead
