+++ backend/src/types/responses/like_list.rs
@@ -1,3 +1,4 @@
+use crate::actions::user::UserRead;
 use crate::domain::user::User;
 use crate::types::*;
 use serde::Serialize;
