+++ backend/src/types/responses/mod.rs
@@ -22,7 +22,6 @@ mod street_stats_response;
 mod tree_list;
 mod tree_list_item;
 mod tree_stats_response;
-mod user_response;
 pub use comment_list::*;
 pub use duplicates_response::*;
 pub use file_status_response::*;
@@ -47,4 +46,3 @@ pub use street_stats_response::*;
 pub use tree_list::*;
 pub use tree_list_item::*;
 pub use tree_stats_response::*;
-pub use user_response::*;
