+++ backend/src/types/responses/single_tree_response.rs
@@ -1,3 +1,4 @@
+use crate::actions::user::UserRead;
 use crate::domain::user::User;
 use crate::types::*;
 use serde::Serialize;
